Calculate association with the reference for the aggregated profile

Calculate association with the reference for the aggregated profile

calAggExpressionCorrelation(
  spe,
  celltype,
  sce_ref,
  ref_celltype = NULL,
  method = c("pearson"),
  spe_exprs_values = "logcounts",
  ref_exprs_values = "mean"
)

Arguments

spe

A SpatialExperiment object

celltype

A string or a vector indicating the cell type information.

sce_ref

A SingleCellExperiment object for the reference data

ref_celltype

A vector with length of the sample indicating the cell type label.

method

method to compute similarity or distance ("pearson" or "cosine"). Default is set to "pearson".

spe_exprs_values

A string indicating which assay of `spe` contains the expression values. Default is set to "logcounts".

ref_exprs_values

A string indicating which assay of `sce_ref` contains the expression values. Default is set to "mean".

Value

A SpatialExperiment object