Calculate the baseline metrics of segmented cell from cell segmentation output

Calculate the baseline metrics of segmented cell from cell segmentation output

calBaselineCellMetrics(
  spe,
  metrics = c("total_transciprts", "total_genes", "total_cells", "meanExprsPct_cells"),
  rerun = TRUE,
  verbose = TRUE
)

Arguments

spe

A Spatial Experiment object

metrics

A character vector indicating the metrics to be calculated

rerun

A logical value indicates whether to recalculate existing metrics.

verbose

A logical value indicates whether to print out running messages.

Value

A SpatialExperiment object