Read Vizgen output

Read Vizgen output

readVizgen(
  data_dir,
  filter_gene_pattern = "Blank",
  tiff_path = NULL,
  method_name = "Vizgen"
)

Arguments

data_dir

a character specifying a single output directory from Vizgen. The data_dir should include one cell_by_gene.csv and one cell_metadata.csv.

filter_gene_pattern

Pattern of the genes to be filtered out.

tiff_path

a character specifying the path of a tiff file output from BIDCell.

method_name

a character specifying the name of the method.

Value

A SpatialExperiment object